rename: 미션안내 cta 카드 파일 이름 수정
