+++ src/app/record/page.tsx
@@ -1,7 +1,7 @@
 import { NavBar } from "@/components/common/NavBar";
 import { Header } from "@/components/dailyRecord/Header";
-import { RecordMissionCard } from "@/components/dailyRecord/MissionCard";
 import { RecordMissionFeed } from "@/components/dailyRecord/MissionFeed";
+import { RecordMissionCard } from "@/components/dailyRecord/MissionInfo";
 import { SharedDiaryCard } from "@/components/dailyRecord/SharedDiaryCard";
 
 import DailyRecordData from "@/mock/dailyRecord.json";
