+++ src/components/dailyRecord/MissionInfo.tsx
